Create ReservationSystem
This project is a small non-graphic software to book, cancel and print the tickets of the journey using a PNR code. It is a booking/billing system which displays you the date, time, airline, destination and all the relevant details according to the input of the user. It also checks for incorrect inputs. 
In the end, it displays the details of the passenger and prints the ticket accordingly.